+++ src/lottie/attributes.js
