+++ src/lottie/edit.js
@@ -1,4 +1,3 @@
-
 /**
  * External dependencies
  */
