+++ src/lottie/editor.scss
